AIO-services/requirements.txt
# Combined requirements for all services
# Core dependencies
numpy==2.2.6
opencv-python-headless==4.12.0.88
pyzmq==27.1.0
python-dotenv==1.1.1

# Roboflow inference SDK (for card and troop detection)
inference-sdk==0.56.0

# HTTP and async support
aiohttp==3.10.11
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
requests==2.32.5
urllib3==2.5.0

# Data processing
attrs==25.3.0
backoff==2.2.1
dataclasses-json==0.6.7
marshmallow==3.26.1
typing-inspect==0.9.0
typing_extensions==4.15.0

# Utilities
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
tqdm==4.67.1
PyYAML==6.0.3
python-dateutil==2.9.0.post0
six==1.17.0

# Visualization and plotting
matplotlib==3.10.6
pillow==11.3.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
pyparsing==3.2.5

# Additional dependencies
defusedxml==0.7.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
propcache==0.3.2
py-cpuinfo==9.0.0
scipy==1.16.2
supervision==0.26.1
yarl==1.20.1
